Add database, application, and attributes
This commit adds a database install to the default recipe. It allows for
both mysql and postgres as database choices. It also adds cloning the
rubycas-server from github into the rubycas user's home directory. And
it adds attributes to support this. It also adds some spec tests to
cover these new features.