Add database and ssl
This commit adds the rubycas user to the rvm group. It also adds the
customized gemfile and an init script for unicorn.

Move database stuff into separate recipe for clarity. Change vm box url
to use vagrantup instead of dropbox. Add data_bags and one data bag item
to be used for authenticator configuration. Add templates for god,
nginx, and rubycas.god config. Remove the unicorn init script since we
are switching to god. Update the application config to use the new
authenticators from data bags. Use our custom fork of the god cookbook
since the current one is not compatible with latest version of god gem.
Add idempotent bootstrap script as an inline shell script to run when
provisioning a vagrant vm. This allows for any customization prior to
chef being run. Currently adds chef 11 gem, vim for development, updates
package lists, and creates a marker file to let us know to exit script
on any subsequent run through. Add nginx recipe to drop off a nginx
config for rubycas. Also install nginx if the attribute is set to true.
Spec needs updated alot.

In order to use rubycas securely, this commit adds to the recipe ssl
related code. It also reverts all of the alignments of equal signs as
this is hard to maintain and results in bad git diffs.

[#21964](https://squaremouth.lighthouseapp.com/projects/63460/tickets/21964)