Separate rubycas from the database
In order to allow rubycas to be installed on a separate server from it's
database, this commit removes the dependencies between the two recipes
so that they can be used independent of each other. In addition to this,
it updates the vagrant file to use two vm's instead of one. This is to
ensure that the database and application recipes can be used completely
separate. And finally, this commit moves configuration for database
connections and ssl into databags. The databag names and items for these
are configured via attributes so that the names can be overridden.

[#21964]