Refactor recipes to be less verbose
In order to keep the recipes concise, this commit refactors the
database, nginx, and default recipes to use less code and still
accomplish the same configuration. This is done by adding a couple
classes to the Rubycas::Helper module that are concerned with data bag
searches for configuration needs. In addition, this commit renames one
of the data bag keys for database configuration from type to database
type. This is because there was a colision with the type method being
used and the value wouldn't be read by the helper classes properly. It
adds a redirect to the nginx template if ssl is enabled so that any non
ssl traffic will get redirected to the ssl connection. And finally it
removes some of the logic in the rubycas god config as it would prevent
the app from starting up after failing to start/restart after 10
minutes.

[#21964]