Move libraries into directory structure
In order to better organize the library for this cookbook, this commit
moves the classes out into their own files and adds directories for
modules. In addition to this, it switches the order that the vm's are
declared in the vagrant file so that the database vm starts prior to the
app server. This is because the database needs to be in place for the
application vm to converge correctly.

[#21964]