Add require_client_recipes and update nginx recipe
In order to provide a method to install client database
libraries/adapters, this commit adds a method to the base class
Rubycas::Helper::Config. It also implements the method in the modules
that get mixed in at runtime. In addition, it reorganizes the nginx
template so that all of the ssl related configuration is done in one if
block. This is for maintainability and conciseness.

[#21964]