Update library directory structure
In order to ensure chef gets all of the necessary files when uploading
this cookbook to a chef server, this commit removes all of the
directories under the library directory. In order to ensure the load
order of these files is preserved, it also adds require statements in
each file that will load their dependencies so we don't have an
undefined constant issue when loading all of the helper classes and
modules.

In addition, this commit also adds a service definition for unicorn so
we can notify it to restart when certain files change. This allows us to
restart unicorn on chef runs that change config files for unicorn.

[#21964](https://squaremouth.lighthouseapp.com/projects/63460-viper/tickets/21964)