Add load balancer support
In order to support rubycas being behind a load balancer, this commit
adds logic to the nginx configuration template for it. In addition, it
adds a new attribute to enable/disable load balancer support and sets it
to false by default. This is because the most generic setup probably
won't be behind a load balancer.

[#23897](https://squaremouth.lighthouseapp.com/projects/63460-viper/tickets/23897)