Fix nginx cookbook dependency
In order to allow the nginx recipe to reliably converge on a node, this
commit targets a specific version of the nginx cookbook. Due to the way
the cookbook breaks compatibility across versions, it is easier to
target a specific version of the cookbook and tailor the nginx recipe to
it. In addition, this commit also adds logic to ensure we get the a
specific version of nginx (1.4.1) installed as part of the recipe as
nginx::source installed something considerably older.

[24181]