Update to latest nginx cookbook
In order to upgrade the rubycas::nginx recipe to the latest nginx
cookbook, this commit adds necessary attributes in the recipe. It sets
attributes that will compile the latest version of nginx as of right now
in their stable line. As well it sets the path to the nginx binary so
that runit can properly start the nginx service.

The last change is to set the ulimit to a more reasonable value in the
Vagrantfile. This is due to a regression bug in vagrant that causes it
to crash during a chef run due to too many files being open at once.

[24181]