Upgrade rubycas cookbook dependencies
In order to enable development to continue on this cookbook, this commit
locks down the versions of test kitchen and chefspec to working versions
that also work with the latest Vagrant and VirtualBox. In addition, this
switches the cookbook to use the mysql-chef_gem cookbook instead of the
now deprecated `mysql::ruby` cookbook.