Update chefspec to version 4.0
In order to keep the test suite up to date for RubyCas, this commit
updates chefspec to version 4.0. It removes the
treat_symbols_as_metadata_keys_with_true_values rspec configuration as
this is now the default behavior. In addition, it removes the be_true
matcher from the nginx_spec as it can't be used with expect syntax.